+++ MisLugares/app/src/main/java/com/example/mislugares/repository/impl/LugaresImpl.java
@@ -182,7 +182,7 @@ public class LugaresImpl implements LugaresRepository {
     }
 
     @Override
-    public void deleteLugar(int id) {
+    public void deleteLugar(long id) {
         String whereClause = COLUMN_ID + " = ?";
         String[] whereArgs = {String.valueOf(id)};
